Rename Style files to Component.Style.tsx instead of componentStyle.tsx
