+++ packages/components/src/components/Avatar/Avatar.style.tsx
