+++ packages/components/src/components/SearchBar/SearchBar.style.tsx
