+++ packages/components/src/index.ts
@@ -2,3 +2,4 @@ export { default as SearchBar } from "./components/SearchBar";
 export { default as Button } from "./components/Button";
 export { default as Video } from "./components/VideoPlayer";
 export { default as Avatar } from "./components/Avatar";
+export { default as Section } from "./components/Section";
